Factor out syscalls for calloc, free, realloc and strdup into dedicated syscalls/{}.c files
