+++ src/syscalls/syscalls.h
@@ -10,8 +10,6 @@
 #ifndef SYSCALLS_H
 #define SYSCALLS_H
 
-#include "enums.h"
-
 char *FTLstrdup(const char *src, const char *file, const char *function, const int line) __attribute__((malloc));
 void *FTLcalloc(size_t n, size_t size, const char *file, const char *function, const int line) __attribute__((malloc)) __attribute__((alloc_size(1,2)));
 void *FTLrealloc(void *ptr_in, size_t size, const char *file, const char *function, const int line) __attribute__((alloc_size(2)));
